requirements.txt
aiogram
langchain-groq
python-dotenv
langchain
langchain_groq
